src/requirements.txt
torch==2.1.1
torchvision==0.16.1
torchtext==0.16.1
pytorch-nlp==0.5.0
numpy==1.26.2
pandas==2.1.4
nltk==3.8.1
seaborn==0.13.0
tqdm==4.66.1
scikit-learn==1.3.2
imbalanced-learn==0.11.0
imbalanced-ensemble==0.2.1
matplotlib==3.8.2
typing-extensions==4.9.0
tflearn==0.5.0